Fix 1.21 provider: Move fetch-images.sh into provision folder (#620)
* Move fetch-images.sh into provision folder

Instead of copying and removing the script we put it directly into the
folder where it belongs. Also we adjust pathes and the README.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Fix verbosity of up.sh colliding with image check

The echoing of what container runtime was used integrated into the list
of images being pulled, which made the image check fail.

Also fixed formatting and some shellcheck warnings.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>