SRIOV provider: Bump nodes image to k8s-1.19 (#621)
* sriov provider: Bump nodes image to k8s-1.19

Currently we use KIND 0.7.0 which dosent provides
k8s-1.19 nodes image, thus its necessary to use newer
version of KIND (e.g: 0.11.1)

On newer versions kubectl binary has been changed.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* kind/common: run the local registry on the cluster network

Since KIND 0.8.0 and above, the cluster will be created
on a dedicated network - "kind", instead of the default
network - "bridge" [1].

This commit changes the network the local registry will
be connected to so it will be the same as the cluster.

[1] https://github.com/kubernetes-sigs/kind/releases/tag/v0.8.0#user-content-breaking-changes

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Rename sriov provider name

Align with other providers name scheme.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Use mirrored nodes image

Signed-off-by: Or Mergi <ormergi@redhat.com>

Co-authored-by: Federico Gimenez <fgimenez@users.noreply.github.com>

Co-authored-by: Federico Gimenez <fgimenez@users.noreply.github.com>