Calico: Configure IP autodetection to use interface regex
Currently, kubevirtci uses Calico's default IP auto-
detection method: first-found.

The disadvantage of this method is that it makes only simplified
guess and on top of that, when new interfaces are added (bridges, veths,
etc.) calico may decide to change the NODE IP, which breaks BGP updates.

This method is not recommended by the documentation either.

It's quite simple to reproduce this, just creating a bridge on a note
and assigning an IP address to it makes calico-node pod to pick up the
assigned address and start using that address for BGP.

Since this is certainly not a desired behaviour, this commit adds
IP_AUTODETECTION_METHOD=interface=eth.*
IP6_AUTODETECTION_METHOD=interface=eth.*
setting to only consider interfaces called eth0, eth1, etc. for IP
address autodetection.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>