Expose CDI upload proxy port in CI clusters
Currently in both kubevirt and CDI our functional tests are using `kubectl port-forward`
to upload the images, which is not very reliable.

Instead, we would like to have a port that is exposed when the kubevirtci cluster is brought up,
that we can look up later and set up a NodePort against.

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>