kind infra, podpreset: Increase PodPreset enabled validation timeout
Since we upgraded kind node image to k8s-1.19, it seems that it takes more time
for changes on kube-apiserver.yaml to propagate (e.g: enabling PodPreset).

Unfortunately there is no API to enable and verify that the PodPreset
feature-gate is enabled.
As for today we check kube-apiserver process command on control-plane node
to validate that the PodPeset feature is enabled.

This workaround is temporary, once [1] will land at kubevirtci we can stop
using PodPreset.

[1] https://github.com/kubernetes-sigs/kind/issues/2318

Signed-off-by: Or Mergi <ormergi@redhat.com>