Add force flag, auto detect latest provision dir
Adds a -f flag that removes existing cdi manifests and replaces them
with the newer ones instead. Also if the provision directory is omitted
the latest provision directory is chosen as default. Together we can now
use it to automate cdi manifest updates.

Split the cdi manifests into two files for operator and custom resource
again so we can install cdi in the usual way. This way we also can omit
the cdi installation from the manifest install as required in
deploy-manifests.sh .

Signed-off-by: Daniel Hiller <dhiller@redhat.com>