Add cdi deployment to kubevirtci clusters (#676)
* Add force flag, auto detect latest provision dir

Adds a -f flag that removes existing cdi manifests and replaces them
with the newer ones instead. Also if the provision directory is omitted
the latest provision directory is chosen as default. Together we can now
use it to automate cdi manifest updates.

Split the cdi manifests into two files for operator and custom resource
again so we can install cdi in the usual way. This way we also can omit
the cdi installation from the manifest install as required in
deploy-manifests.sh .

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add optional cdi deployment to clusters

Adds support for environment variable KUBEVIRT_DEPLOY_CDI that triggers
a deployment of bundled cdi if set to `true`.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add usage implementation

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Patch ip-reconciler cronjob to directly delete the job pods

The check-provision presubmit sometimes flakes because it may see
pods from the whereabouts ip-reconciler cron job that are in
Completed status.

Therefore we patch the job to directly delete the job pods so they do
not interfere with the pod check.

Also we remove the until look from the cdi pod condition check and
increase the timeout.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>