Improve PodPreset enablement validation (#683)
* podpreset: Increase validation timeout

It seems that when kube-apiserver is under pressure it takes
more than 60 seconds for new configurations to propagate.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* podpreset: Wait for kube-system pods restart

PodPreset is enabled changing by changing kube-apiserver pod
manifest, which when changed triggers kube-system pods to restart
along with the kube-apiserver pod.

We should make sure that the cluster is ready after enabling
PodPreset before the cluster-up process continues.

Signed-off-by: Or Mergi <ormergi@redhat.com>