vgpu, Remove podpreset (#708)
Live migration is not supported yet on vgpu.
and the lane assumes there is only one node.

Podpreset is flaky a bit, so better to remove it
since its not needed.

Once it will be needed we can use k8s-1.22 with new kind
version that supports unique uuid built-in.

Signed-off-by: Or Shoval <oshoval@redhat.com>