KinD providers: Mount /var/log/audit to cluster nodes
On Kubevirt test suite we dump various cluster object
stats, guest VM and cluster nodes logs, including auditd logs.
Currently we can't dump the audit log for KinD providers
since the cluster nodes are containers (no auditd).

In order to have audit log on KinD providers as well it
is necessary to mount the host /var/log/audit to the
cluster nodes.

Then on a test failure we will get the audit log for
the time the test were running.

Signed-off-by: Or Mergi <ormergi@redhat.com>