KinD providers: Mount audit logs as read-only (#724)
On Kind clusters auditing at the node level is not supported.
We mount the host audit logs dir in order to enable Kubevirt test
suite reporter to dump them the logs at the time a test fails.

The mount is misconfigured and should be mounted as read-only
to prevent logs corruption.

Signed-off-by: Or Mergi <ormergi@redhat.com>