provision: Remove fluentd logging feature
The logging by fluentd was an attempt to stream logs out
before we had our k8s dump reporter in kubevirt.
k8s reporter is good enough for kubevirt,
and the logging mechanism was not maintained, nor used since.

This commit removes the feature, and as such removes also
the inclusive language used by it.

Signed-off-by: Or Shoval <oshoval@redhat.com>