console: Use screen instead of socat
To connect the nodes using the qemu pty console we were using socat, but
socat is very basic and it does not handle well the shell input. This
change replace socat with screen to have a proper shell over the pty and
also alow you to open multiple panes on it.

Signed-off-by: Quique Llorente <ellorent@redhat.com>