cluster-up: Use podman.socket's default path
These changes set the path of the podman's socket to
'${XDG_RUNTIME_DIR}/podman/podman.socket' so podman can be started with
'systemctl start --user podman.socket' which is the standard way of
starting podman as user on distributions using systemd.

Also silence stderr output of `docker ps` command so behavior is uniform.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>