cluster-up: Avoid a known conmon issue when using podman (#743)
There is currently a known conmon issue [1] when podman based containers
dump large amounts of data to stdout. gocli would previously do this
when copying the kubectl binary to the host machine.

This change replaces that with a simple volume mount workaround when
using podman until the underlying conmon issue is resolved.

[1] https://github.com/containers/conmon/issues/315

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>