cluster-up: Unify usage of rootless or rootful podman
This adds the '--remote' and '--url' argument to every podman call to
unify the usage of rootless or rootful podman. By using the same socket
everywhere it is guaranteed that rootless and rootful podman are not mixed.
Prior providing a socket to a rootful podman would still leave some
instances where a rootless podman (without socket) was used.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>