vm.sh: Add iptables DNAT rules for rootless podman (#756)
This adds iptables DNAT rules for rootless podman. Because traffic is
originating from the loopback adapter when using rootles podman /
slirp4netns, the existing iptables rules did not apply and
'make cluster-up' was failing on rootless podman. This patch checks
whether a rootless container is used and adds the corresponding
iptables rules.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>