k8s-1.22: Support provisioning DualStack and SingleStack IPv6 (#806)
* k8s-1.22: Support provisioning DualStack and Single Stack IPv6

In order to ease maintaining providers,
support both modes in a single provider.
This way when we present a new provider, we will able
to easily create also an IPv6 only variant of it.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* k8s-1.22-ipv6: Align code to be the same of k8s-1.22

Code supports now both DualStack and Single Stack IPv6,
according an argument.

Align the two folders to have the same code.
The main provision.sh will auto detect which settings
it should use according the folder name.

Signed-off-by: Or Shoval <oshoval@redhat.com>