Update cluster provisioning image to fedora 36 (#799)
* Update cluster provisioning image to fedora 36

Fedora 31 is no longer supported so we should update to use Fedora 36.
This change also includes a bump to the dockerize version to the latest
release.

Signed-off-by: Brian Carey <bcarey@redhat.com>

* Update qemu-img command to include backing file format

The backing file format is required since qemu v6.1[1]

[1] https://github.com/qemu/qemu/blob/master/docs/about/removed-features.rst#qemu-img-backing-file-without-format-removed-in-61

Signed-off-by: Brian Carey <bcarey@redhat.com>

* Create manifest directories ahead of using scp to copy files

The manifest directories have to be in place before carrying out the
copy with scp.

If the directories are not in place, the scp command will fail due to
not being able to find the directories.

Signed-off-by: Brian Carey <bcarey@redhat.com>