podman: Support non root user via root socket (#824)
* podman: Support non root user via root socket

In order to use roootful podman by a non root user,
podman supports a socket (like docker does).

Adapt code to support it.
Changes:
1. Support the default root socket path /run/podman/podman.socket
2. Download the kubectl according the cluster version instead
copying it from the node.
We copy it from the node as a workaround of podman issue,
but it is copied with root only permissions, and then the non root
user can't access it.

Note that kubectl is needed to be presented, it will used only
for checking the cluster version and then the kubectl with the same
version as the cluster will be downloaded and used (as before).

Signed-off-by: Or Shoval <oshoval@redhat.com>

* vm based providers: Cosmetics

Rename _docker_socket to _cri_socket
as it can be either podman or docker socket.

Signed-off-by: Or Shoval <oshoval@redhat.com>