Bump SR-IOV proivider components (#821)
* sriov provider: Bump sriov-device-plugin to v3.4

The manifest is also updated according to the manifest
from v3.4.0 release [1].
Image tag is changed explicitly to v3.4 because the original
manifest specify an older version.

[1] https://github.com/k8snetworkplumbingwg/sriov-network-device-plugin/blob/v3.4.0/deployments/k8s-v1.16/sriovdp-daemonset.yaml

Signed-off-by: Or Mergi <ormergi@redhat.com>

* sriov provider: Bump sriov-cni to v2.6.2

The manifest is also updated according to the manifest
from release [1].
The tag is changed v2.6.2 explicitly because the original
manifest specify an older version.

[1] https://github.com/k8snetworkplumbingwg/sriov-cni/blob/v2.6.2/images/k8s-v1.16/sriov-cni-daemonset.yaml

Signed-off-by: Or Mergi <ormergi@redhat.com>

* sriov provider: Bump Multus to v3.8

The manifest is also updated according to the manifest
from release [1].

The new manifest include ConfigMap for injecting configuration
to Multus, it is removed because on SR-IOV provider the auto
generated config is used.

[1] https://github.com/k8snetworkplumbingwg/multus-cni/blob/v3.8/images/multus-daemonset.yml

Signed-off-by: Or Mergi <ormergi@redhat.com>