provision: Copy manifests only on linux phase
Currently manifests are copied both on linux
and k8s phase.
In case using phased mode, the changes
that are done on the linux phase would be overridden.
Some changes should be done on linux phase
because the images are pre pulled on this phase,
so it should be done before pre pulling.

Additional change:
Verify that the script exists just before running it.

Signed-off-by: Or Shoval <oshoval@redhat.com>