vm based providers: Expose a UDP port for DNS (#867)
* base image: Add debug packages

Those packages add helpfull commands
such as ping, nslookup, tcpdump,
that are needed when developing network features.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* vm based providers: Expose a UDP port for DNS

Expose a custom UDP port (selected 31111)
by the vm based providers.
This allows to expose a nodePort of a service,
set it's node port to 31111,
and then it will be reachable from the host itself.

Signed-off-by: Or Shoval <oshoval@redhat.com>

Signed-off-by: Or Shoval <oshoval@redhat.com>