New kubevirt k8s 1.26 rc provider (#907)
* Coping necopying necessary directories
Currenly just coping k8s 1.25 files

Signed-off-by: bmordeha <bmodeha@redhat.com>

* Update version to be rc version of 1.26
Because 1.26 doesn't have release version right now

Signed-off-by: bmordeha <bmodeha@redhat.com>

* Using previous version of kubectl and kubelet
When Kubernetes version is rc/alpha/beta version
the same release isn't available for these packages.
According to:
https://kubernetes.io/docs/setup/production-environment/tools/kubeadm/create-cluster-kubeadm/#kubeadm-s-skew-against-the-kubernetes-version
kubeadm can be used with Kubernetes components
that are the same version as kubeadm or one version older.
Therefore i had to switch the binary of kubeadm
to use the rc version.
The reason we install previous version of kubeadm
using dnf is for the dependencies packages.

Signed-off-by: bmordeha <bmodeha@redhat.com>

* Moving default of k8s 1.26 rc vesion to centos9

Signed-off-by: bmordeha <bmodeha@redhat.com>

* Add Kubenetes version to the logs
Get the Kubernetes version of the cluster control
plane with the kubectl version --short command.

Signed-off-by: bmordeha <bmodeha@redhat.com>

Signed-off-by: bmordeha <bmodeha@redhat.com>
Co-authored-by: bmordeha <bmodeha@redhat.com>