vm based cluster: Support both dual stack and single stack
In order to reduce busywork and number of providers,
Update the provider to support both single stack and dual stack
according cluster-up configuration.

This commit introduces KUBEVIRT_SINGLE_STACK which determines
which stack to use (default: dual stack).

Remove the logic that allows provisioning single stack,
as we don't need it, and it will be easier to maintain without it.

Signed-off-by: Or Shoval <oshoval@redhat.com>