vm based providers: Introduce robust stack provider (#954)
* psa: Add missing psa configuration for kubeadm_ipv6.conf

Signed-off-by: Or Shoval <oshoval@redhat.com>

* k8s-1.25-ipv6: Delete provider

Create robust stack provider.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* publish: Create only the robust provider

Signed-off-by: Or Shoval <oshoval@redhat.com>

* vm based cluster: Support both dual stack and single stack

In order to reduce busywork and number of providers,
Update the provider to support both single stack and dual stack
according cluster-up configuration.

This commit introduces KUBEVIRT_SINGLE_STACK which determines
which stack to use (default: dual stack).

Remove the logic that allows provisioning single stack,
as we don't need it, and it will be easier to maintain without it.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* single stack: Move single stack sanity check

Use the sanity check as part of cluster-up
Fix it to wait for the calico pod to be ready before
checking its IP to avoid flakes.

Run single stack cluser-up sanity check as part
of check cluster-up.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* gocli: Remove provision stack type argument

Since we don't support anymore provisioning of single stack
remove it's flag.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* k8s-1.26: Support robust stack provider

Signed-off-by: Or Shoval <oshoval@redhat.com>

* k8s-1.26-centos9: Adapt to robust provider

Signed-off-by: Or Shoval <oshoval@redhat.com>

---------

Signed-off-by: Or Shoval <oshoval@redhat.com>