check cluster-up: Verify Multus
Check that Multus is functional by create a pod with
secondary network and check if its Ready.
The pod wont reach Ready state if Multus is broken in a way
and fail early on init phase.

Kubevirt CI check-provision lanes runs the check-cluster-up.sh script.
Checking that Multus is functional as part of check-provision lanes
prevent merging changes that will turn to be breaking only when they
land at kubevirt/kubevirt.

Signed-off-by: Or Mergi <ormergi@redhat.com>