kind, vgpu: Bump vgpu kind to k8s-1.25 (#979)
* kind: Bump vgpu kind to k8s-1.25

Do not config cpu manager for vgpu,
because kind 1.24+ has this bug for cpu manager:
https://github.com/kubernetes-sigs/kind/issues/2999

Since we don't use cpu manager on vgpu lane we can bump to k8s-1.25
and remove cpu manager.

Rename lane.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* kind: Rename functions

The functions can add extra mounts / cpu manager to non worker nodes,
it depends where it is called.
If it is called before the worker snippet in the manifest
it will configure it for the control-plane, otherwise
for the worker node.

Rename it to reflect it.

Signed-off-by: Or Shoval <oshoval@redhat.com>

---------

Signed-off-by: Or Shoval <oshoval@redhat.com>