provision-manager: Switch to fedora
It is possible that gocli Alpine is the cause for the hangs
during provision we see lately.
It started with the pman PR, and also running pman itself hanged once.
Lets move to fedora, to see if it is related.
We can also consider using go-git and then we can remove the OS totally.

Signed-off-by: Or Shoval <oshoval@redhat.com>