Add container stop to ensure it stops after shutdown
There are occassions where the container fails to stop after the
shutdown of the VM is issued. This causes the job to time out waiting
for the container to stop[1].

Add a step to stop the container after the shutdown has been issued to
ensure that the container is stopped.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_kubevirtci/1018/check-provision-k8s-1.26-centos9/1679560587169763328#1:build-log.txt%3A4865

Signed-off-by: Brian Carey <bcarey@redhat.com>