kind: Auto detect kubectl path
Once in a while the kubectl path in kind node
is changed (for example if the base image is changed).

So far it was only one of two paths, and we are using it
only for copying kubectl from the node.
Auto detect it, so we won't need to change it whenever it changes.

Signed-off-by: Or Shoval <oshoval@redhat.com>