kind, sriov: Introduce kind sriov bump, generic provider name, and bump to k8s-1.28 (#1122)
* kind: Introduce bump-kind-sriov script

The script eases bumping kind-sriov provider.
Later we can use it for automation.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* kind: Auto detect kubectl path

Once in a while the kubectl path in kind node
is changed (for example if the base image is changed).

So far it was only one of two paths, and we are using it
only for copying kubectl from the node.
Auto detect it, so we won't need to change it whenever it changes.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* kind-sriov: Duplicate kind-1.27-sriov folder

Use kind-sriov generic name, it will allow easier updating
of the CI jobs once we update the provider.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* kind: Extract parameters to files

Allows updating them in more elegant way,
without changing source files themselves.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* kind-sriov: Bump to v0.19.0, k8s-1.28

Signed-off-by: Or Shoval <oshoval@redhat.com>

* kind: Fix bash error

Error is because using single brackets requires quoting.

cluster/kind/common.sh: line 303: [: !=: unary operator expected

Signed-off-by: Or Shoval <oshoval@redhat.com>

---------

Signed-off-by: Or Shoval <oshoval@redhat.com>