Build kubevirtci nodes from static centos stream base image (#1140)
* Update kubernetes provisioning to use static centos base image

Signed-off-by: Brian Carey <bcarey@redhat.com>

* Move kubernetes specific steps to k8s_provision.sh

Move cri-o and kubernetes installs to the k8s_provision.sh script

provision.sh to handle the remaining install of rpm dependencies

This should allow publishing a centos stream base image with all of
the required rpm dependencies included.

Signed-off-by: Brian Carey <bcarey@redhat.com>

* Update publish.sh to allow pushing linux base separately

The publish for the kubevirtci nodes is updated to build from a static
centos stream 9 base.

A periodic job will use this publish script to push updated centos 9
base images to quay and create a PR with the updated base image so that
tests can run against it.

Signed-off-by: Brian Carey <bcarey@redhat.com>

* gocli: Remove deprecated ioutil ReadFile function

Signed-off-by: Brian Carey <bcarey@redhat.com>

* centos: update image suffix to base rather than dev

Signed-off-by: Brian Carey <bcarey@redhat.com>

---------

Signed-off-by: Brian Carey <bcarey@redhat.com>