gocli,k8s: Copy provider scripts during k8s phase to ensure up to date (#1148)
The scripts are initially copied when building the linux base image but
these maybe out of date if the linux base has not been updated in some
time.

Also changes to these scripts won't be tested in the presubmit lanes as
the updated scripts won't be available in the linux base.

Copy the provider scripts again during the k8s phase to ensure that they
are consistent with main

Signed-off-by: Brian Carey <bcarey@redhat.com>