Introduce KUBEVIRT_CPU_MANAGER_POLICY
This change introduces the KUBEVIRT_CPU_MANAGER_POLICY env var that
allows a user to enable the CPUManager with a specific policy.

The static policy is the only supported policy at present with the
full-pcpus-only option provided.

Two FIXMEs are left in the code to move to config drop-ins and fix an
issue with HereDocs that could clean up this implementation slightly.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>