Introduce `KUBEVIRT_NUM_NUMA_NODES`, `KUBEVIRT_NUM_VCPU` and `KUBEVIRT_CPU_MANAGER_POLICY` (#1171)
* Introduce KUBEVIRT_NUM_NUMA_NODES

This change introduces the KUBEVIRT_NUM_NUMA_NODES env variable allowing
for 2 or more NUMA nodes to be defined per host. This requires that both
the amount of memory and vCPUs be divisible by this value in order for
an even distribution of resources across the resulting nodes.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>

* Introduce KUBEVIRT_NUM_VCPU

This change introduces the KUBEVIRT_NUM_VCPU env variable allowing users
to define a number of vCPUS exposed as cores for each host.

This was previously hardcoded to 6 so the value is retained as the
default to avoid breaking any existing envs where this value is assumed.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>

* Introduce KUBEVIRT_CPU_MANAGER_POLICY

This change introduces the KUBEVIRT_CPU_MANAGER_POLICY env var that
allows a user to enable the CPUManager with a specific policy.

The static policy is the only supported policy at present with the
full-pcpus-only option provided.

Two FIXMEs are left in the code to move to config drop-ins and fix an
issue with HereDocs that could clean up this implementation slightly.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>

* gocli: Replace NewEnvClient with NewClientWithOpts

NewEnvClient being deprecated in favor of NewClientWithOpts:

https://pkg.go.dev/github.com/docker/docker/client#NewEnvClient

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>

---------

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>