Creates the utils package
Contains the K8sDynamicClient type which will be used to create an arbitrary object from a manifest. as well as a test client based on K8s Fake. In addition to the SSH Client type which contains interfaces to connect to nodes in the node network through the forwarded ssh port by the Dnsmasq container

Signed-off-by: aerosouund <aerosound161@gmail.com>