Introduce k8s Opts
K8s opts map to the scripts that create a kubernetes deployment of something on the CI cluster. The opts depend on the k8s client type to apply a list of manifests and leverage the test client for testing while passing reactors to the test client to configure necessary behavior. the embed package is used to ship the necessary manifests with those opts. This includes rook-ceph, nfs-csi, cdi, aaq, multus, cnao and prometheus

Signed-off-by: aerosouund <aerosound161@gmail.com>