Creates k8s base provider
The k8s base provider is an abstraction that makes it easier to deal with a running provider. it leverages the Opts in provisioning the cluster. Some of its logic is a refactor of the run method and the other functions that were in run.go but didn't have a type attached to them. the Flagmap variable serves as a mapping between each flag and its corresponding OptFunc that sets the field in the struct

Signed-off-by: aerosouund <aerosound161@gmail.com>