Rewrite run to use the KubevirtProvider type
The logic for run has been moved from running directly in the function to the provider. run was changed to parse the flags based on their types, initialize a provider and run its Start method. extra flags have been added for the options that previously were entirely in the bash code. a default container suffix has been introduced as none existed before. and unless passed provisioning would fail. as well as change the default for background and dns port to match what the bash code always set as default in k8s_provider_common.sh

Signed-off-by: aerosouund <aerosound161@gmail.com>