feat: Kind base provider
The base provider is a valid provider that can be run standalone for a stock kind cluster, And is also the base for
any other type of kind cluster.
It is container runtime agnostic and depends on the sigs.k8s.io/kind/pkg/cluster package which is introduced in this commit as a dependency.
It uses this package to provision a kind cluster and provides extra functionality by running a registry container and configuring the nodes
to reach it.

Signed-off-by: aerosouund <aerosound161@gmail.com>